+++ hdbscan/hdbscan_.py
@@ -229,12 +229,12 @@ def _hdbscan_prims_kdtree(
     gen_min_span_tree=False,
     **kwargs
 ):
-    if X.dtype != np.float64:
-        X = X.astype(np.float64)
+    if X.dtype != np.float32:
+        X = X.astype(np.float32)
 
     # The Cython routines used require contiguous arrays
     if not X.flags["C_CONTIGUOUS"]:
-        X = np.array(X, dtype=np.double, order="C")
+        X = np.array(X, dtype=np.float32, order="C")
 
     tree = KDTree(X, metric=metric, leaf_size=leaf_size, **kwargs)
 
@@ -271,12 +271,12 @@ def _hdbscan_prims_balltree(
     gen_min_span_tree=False,
     **kwargs
 ):
-    if X.dtype != np.float64:
-        X = X.astype(np.float64)
+    if X.dtype != np.float32:
+        X = X.astype(np.float32)
 
     # The Cython routines used require contiguous arrays
     if not X.flags["C_CONTIGUOUS"]:
-        X = np.array(X, dtype=np.double, order="C")
+        X = np.array(X, dtype=np.float32, order="C")
 
     tree = BallTree(X, metric=metric, leaf_size=leaf_size, **kwargs)
 
@@ -318,8 +318,8 @@ def _hdbscan_boruvka_kdtree(
     if core_dist_n_jobs < 1:
         core_dist_n_jobs = max(cpu_count() + 1 + core_dist_n_jobs, 1)
 
-    if X.dtype != np.float64:
-        X = X.astype(np.float64)
+    if X.dtype != np.float32:
+        X = X.astype(np.float32)
 
     tree = KDTree(X, metric=metric, leaf_size=leaf_size, **kwargs)
     alg = KDTreeBoruvkaAlgorithm(
@@ -362,8 +362,8 @@ def _hdbscan_boruvka_balltree(
     if core_dist_n_jobs < 1:
         core_dist_n_jobs = max(cpu_count() + 1 + core_dist_n_jobs, 1)
 
-    if X.dtype != np.float64:
-        X = X.astype(np.float64)
+    if X.dtype != np.float32:
+        X = X.astype(np.float32)
 
     tree = BallTree(X, metric=metric, leaf_size=leaf_size, **kwargs)
     alg = BallTreeBoruvkaAlgorithm(
